Split the remaining commands_test.go into an import_test.go and a root_test.go
root_test.go contains code shared between the test code for commands.
